version 0.60

+++ src/Meta.h
@@ -10,7 +10,7 @@
 *************************************************************/
 int	Init_MatFile_Read(int NumPop);
 int	Open_NewMatFile_Read(int idx, const char * FileName);
-int	Mat_GetData(int idx, double * mat, int start, int nmarker);
+int	Mat_GetData(int idx, double * mat, long start, int nmarker);
 int Close_MatFile(int idx);
 
 /*************************************************************
@@ -23,7 +23,7 @@ int	Open_NewMatFile_Save(const char * FileName);
 int	Mat_PutData(double * mat, int size);
 int	Mat_Check_Saved();
 int	Mat_Num_Sets();
-int	Mat_GetStart_Pos(int * pos, int * size);
+int	Mat_GetStart_Pos(double * pos, int * size);
 int	Close_Write_MatFile();
 
 /*************************************************************
